+++ core/events.lua
@@ -166,17 +166,6 @@ local M = {}
 --     reported as *both* "Control" and "Alt" due to a Scintilla limitation with GTK.
 -- @field INITIALIZED (string)
 --   Emitted after Textadept finishes initializing.
--- @field KEYPRESS (string)
---   Emitted when pressing a key.
---   If any handler returns `true`, the key is not inserted into the buffer.
---   Arguments:
---
---   * _`code`_: The numeric key code.
---   * _`shift`_: The "Shift" modifier key is held down.
---   * _`ctrl`_: The "Control" modifier key is held down.
---   * _`alt`_: The "Alt"/"Option" modifier key is held down.
---   * _`cmd`_: The "Command" modifier key on macOS is held down.
---   * _`caps_lock`_: The "Caps Lock" modifier is on (GTK only).
 -- @field LEXER_LOADED (string)
 --   Emitted after loading a language lexer.
 --   This is useful for overriding a language module's key bindings or other properties since
@@ -266,10 +255,11 @@ local M = {}
 --   * _`index`_: The numeric index of the clicked tab.
 --   * _`button`_: The mouse button number that was clicked, either `1` (left button), `2`
 --     (middle button), `3` (right button), `4` (wheel up), or `5` (wheel down).
---   * _`shift`_: The "Shift" modifier key is held down.
---   * _`ctrl`_: The "Control" modifier key is held down.
---   * _`alt`_: The "Alt"/"Option" modifier key is held down.
---   * _`cmd`_: The "Command" modifier key on macOS is held down.
+--   * _`modifiers`_: A bit-mask of any modifier keys held down: `view.MOD_CTRL`,
+--     `view.MOD_SHIFT`, `view.MOD_ALT`, and `view.MOD_META`. On macOS, the Command modifier
+--     key is reported as `view.MOD_CTRL` and Ctrl is `view.MOD_META`. Note: If you set
+--     `view.rectangular_selection_modifier` to `view.MOD_CTRL`, the "Control" modifier is
+--     reported as *both* "Control" and "Alt" due to a Scintilla limitation with GTK.
 -- @field TAB_CLOSE_CLICKED (string)
 --   Emitted when the user clicks a buffer tab's close button.
 --   When connecting to this event, connect with an index of 1 if the handler needs to run